manage user rights

+++ routes/ad.routes.js
@@ -34,7 +34,8 @@ router.post(
             return;
         }
         res.json({ fileUrl: req.file.path });
-});
+    }
+);
 
 router.post('/', isAuthenticated, (req, res) => {
     const {
@@ -64,12 +65,22 @@ router.post('/', isAuthenticated, (req, res) => {
     })
         .then((newAd) => {
             console.log('newAd', newAd);
-             User.findByIdAndUpdate({_id: owner}, {
-                $push: { ads: newAd._id},
-            }).exec()
+            User.findByIdAndUpdate(
+                { _id: owner },
+                {
+                    $push: { ads: newAd._id },
+                }
+            ).exec()
+            return newAd
         })
-        .then(() => res.status(201).json({ message: 'ad has successfully been created' }))
-        .catch(() => res.status(500).json({ message: 'error when creating the ad' }));
+        .then((newAd) =>
+            res
+                .status(201)
+                .json({ message: 'ad has successfully been created', _id: newAd._id })
+        )
+        .catch(() =>
+            res.status(500).json({ message: 'error when creating the ad' })
+        );
 });
 
 router.get('/:adId', (req, res) => {
@@ -85,7 +96,19 @@ router.get('/:adId', (req, res) => {
         .catch((error) => res.json(error));
 });
 
-router.put('/:adId/edit', isAuthenticated, (req, res) => {
+const isOwner = async (req) => {
+    const { adId } = req.params;
+    const userId = req.payload._id;
+    
+    return await Ad.findById(adId)
+        .then((ad) => {
+            console.log('userId', userId)
+            console.log('ad owner', String(ad.owner))
+            return String(ad.owner) === userId
+        }) 
+};
+
+router.put('/:adId/edit', isAuthenticated, async (req, res, next) => {
     const { adId } = req.params;
 
     if (!mongoose.Types.ObjectId.isValid(adId)) {
@@ -93,13 +116,17 @@ router.put('/:adId/edit', isAuthenticated, (req, res) => {
         return;
     }
 
-    Ad.findByIdAndUpdate(adId, req.body, { new: true })
-        .populate('owner')
-        .then((ad) => res.status(200).json(ad))
-        .catch((error) => res.json(error));
+    if(await isOwner(req)) {
+        Ad.findByIdAndUpdate(adId, req.body, { new: true })
+                    .populate('owner')
+                    .then((ad) => res.status(200).json(ad))
+                    .catch((error) => res.json(error));
+    } else {
+        res.redirect(303, '/ads')
+    }
 });
 
-router.delete('/:adId', isAuthenticated, (req, res) => {
+router.delete('/:adId', isAuthenticated, async (req, res) => {
     const { adId } = req.params;
 
     if (!mongoose.Types.ObjectId.isValid(adId)) {
@@ -107,13 +134,15 @@ router.delete('/:adId', isAuthenticated, (req, res) => {
         return;
     }
 
-    Ad.findByIdAndRemove(adId)
-        .then(() =>
-            res.json({
-                message: `Ad with ${adId} is removed successfully.`,
-            })
-        )
-        .catch((error) => res.json(error));
+    if(await isOwner(req)) {
+        Ad.findByIdAndRemove(adId)
+                .then(() =>
+                    res.json({message: `Ad with ${adId} is removed successfully.`})
+                )
+                .catch((error) => res.json(error));
+    } else {
+        res.redirect(303, '/ads')
+    }
 });
 
 module.exports = router;
